mutating and shuffle
