+++ .github/mut_config.yml
@@ -1,0 +1,6 @@
+skip_without_test: false
+skip_with_build_tags: false
+json_output: true
+silent_mode: false
+exclude_dirs:
+  - messages/proto
